Rework sizing of canvas and fix focus issue
This commit reworks the layout of the HTML to better utilize all the
available space in the browser window, and avoids hard-coding the space
to reserve for the controls in the Javascript. It also now allows the
canvas to be rectangular to increase the viewable area, and fixes a
problem with focusing on the config panel on iOS devices.